Use host nginx for phpmyadmin service
Don't expose any ports for the phpmyadmin service. Instead, we'll
use the host nginx to serve port 8081 to this container, which will
allow us to use the host's SSL certificates.

This is much easier than hacking the phpmyadmin image to make its
internal nginx server use SSL.

Also: make the config read-only (because it should be).