Add nginx service
We want to transition fully to containerized services, so we need
to start migrating some of the services that are currently served
by the host nginx.

One big benefit of this change is that the containerized nginx
can refer to other containerized services by the docker service
name instead of hard-coding a dynamic IP address (e.g. we can
use `http://phpmyadmin` instead of `http://172.18.0.3`).

The first service we are migrating is phpmyadmin.