phpmyadmin: remove config file
Now that we are no longer using the host mysql, we can use the
standard docker network link to the mysql container instead, which
requires no special configuration file.

We are also performing the IP address blocking at the Nginx level,
so that is not necessary either (and it was burdensome to implement
as well, since the `proxy_pass` of Nginx caused the wrong IP to be
sent to the phpmyadmin container).

With no other config options to specify, we remove the config file.