Use nginx service as reverse proxy for smr
As we migrate to the new host server, we will need to use the nginx
server as the reverse proxy for the main smr website instead of the
system nginx (since there will not be one).

We also add the certbot service for certificate renewal.

This particular setup prevents us from using awstats, but we can
either a) forego awstats in favor of simply tracking login page
statistics with Google Analytics, or b) trying to configure an
awstats docker image.