nginx/smr.conf: Redirect to www.smrealms.de
Use www.smrealms.de as the canonical server name, and send 301
redirects to it from any naked smrealms.de requests.

This definitely solves issues we've had with Facebook validation:

> FacebookSDKException: Cross-site request forgery validation failed.
> Required param "state" missing from persistent data.

It may also help with the persistent issue we've seen with attempting
to INSERT duplicate `session_id` rows into `active_session`, but it
is not yet clear if this is related to www vs. non-www.