Workaround to avoid having the warning message in the dump file
When mysql is used with -u... -p... a warning message is printed to
standard out. That message is then piped to the dump file, which makes
it non sql conform. The password is available on the host in .env and
inside the container as environment variable. Simply use that to create
a config file that is used by any client connections.