docker-compose.yml: restart unless-stopped
Since the nginx service depends on both the smr and phpmyadmin
services, add phpmyadmin as a dependency to smr. Then, adding
`restart: unless-stopped` to all these services accomplishes
two things:

1. It ensures that all essential smr services are online unless
   we have explicitly stopped them, which reduces the likelihood
   of service disruption.

2. It solves (in an unsatisfying way) the problem of nginx failing
   to start unless the smr and phpmyadmin services are already started.

For 2, the issue is that there is a degeneracy between logical
dependencies and functional dependencies in docker-compose. The smr
service functionally depends on the nginx service (which needs to
be running for the smr server to be functional). However,
the nginx service logically depends on the smr and phpmyadmin services
(since the references to the phpmyadmin and smr services in nginx.conf
causes an error unless these services are already exposed on the local
network).

However, specifying both the logical and functional dependencies
would result in a "circular dependencies" error from docker-compose.

A workaround is to abuse the "restart: unless-stopped" condition
so that, even though nginx will fail on its first try to start,
it will succeed a little later as the phpmyadmin and smr services
spin up afterwards. This allows us to specify only the functional
dependencies. It is unsatisfying because the nginx container will
report errors until it successfully starts up.

The alternative is to specify the logical dependencies instead, but
then you must remember to specify all services on the command-line
in order for the services to be functional, i.e.
`docker-compose up --build -d nginx phpmyadmin smr`.

See https://github.com/docker/compose/issues/2589 for a relevant
feature request.