Restrict smr service to www.smrealms.de
Without a "Host" traefik.frontend.rule for the smr service, if any
of the auxiliary services with non-www subdomains (e.g. api, beta,
wiki) are not up, then they will redirect to www.smrealms.de, which
we do not want to happen.

By adding "Host:www.smrealms.de", we make sure you only get to the
game if you are using the game's URL.

(Note that because we have the non-www to www redirect, the bare
smrealms.de URL will still redirect to www.smrealms.de.)