get_live_version.sh: fix issue with traefik conflicts
We cannot run `docker-compose` to get the latest image version
because it will spin up a temporary container that will confuse
traefik (because it will have all the traefik labels).

If we instead run `docker` directly, then we avoid all of the
orchestration overhead and interact only with the image itself.

This fixes "Bad Gateway" errors that occur after this script
is run.