docker-compose.yml: fix implicit boolean conversion
The line `restart: no` results in the following error:

services.npc.restart contains an invalid type, it should be a string

This is because in YAML, an un-quoted `no` is implicitly converted
into a boolean. We need to quote it to force YAML to parse it as
a string.

This is "working as intended" according to the Docker team. See
https://github.com/docker/compose/issues/3672.