docker-compose.yml: add logging limits to traefik
Over the past 2 months, the live traefik service has generated ~85MB
of logs each day. By default, Docker does not trim log files, so for
a long-running service this eventually becomes a problem.

We limit this to 7 100MB logfiles, which should give us at least a
week's worth of logs, even at times of high player activity.

The rotated logs will be compressed, so if this causes any performance
issues with `docker logs`, then that setting can be re-evaluated.